+++ src/anemoi/graphs/generate/tri_icosahedron.py
@@ -5,48 +5,54 @@ from collections.abc import Iterable
 import networkx as nx
 import numpy as np
 import trimesh
-from sklearn.metrics.pairwise import haversine_distances
 from sklearn.neighbors import BallTree
 
+from anemoi.graphs.generate.masks import KNNAreaMaskBuilder
 from anemoi.graphs.generate.transforms import cartesian_to_latlon_rad
+from anemoi.graphs.generate.utils import get_coordinates_ordering
 
 
-def create_icosahedral_nodes(
-    resolutions: list[int],
+def create_tri_nodes(
+    resolution: int, aoi_mask_builder: KNNAreaMaskBuilder | None = None
 ) -> tuple[nx.DiGraph, np.ndarray, list[int]]:
-    """Creates a global mesh following AIFS strategy.
+    """Creates a global mesh from a refined icosahedron.
 
     This method relies on the trimesh python library.
 
     Parameters
     ----------
-    resolutions : list[int]
-        Levels of mesh resolution to consider.
+    resolution : int
+        Level of mesh resolution to consider.
+    aoi_mask_builder : KNNAreaMaskBuilder
+        KNNAreaMaskBuilder with the cloud of points to limit the mesh area, by default None.
 
     Returns
     -------
     graph : networkx.Graph
-        The specified graph (nodes & edges).
+        The specified graph (only nodes) sorted by latitude and longitude.
     coords_rad : np.ndarray
         The node coordinates (not ordered) in radians.
     node_ordering : list[int]
-        Order of the nodes in the graph to be sorted by latitude and longitude.
+        Order of the node coordinates to be sorted by latitude and longitude.
     """
-    sphere = trimesh.creation.icosphere(subdivisions=resolutions[-1], radius=1.0)
+    sphere = trimesh.creation.icosphere(subdivisions=resolution, radius=1.0)
 
     coords_rad = cartesian_to_latlon_rad(sphere.vertices)
 
-    node_ordering = get_node_ordering(coords_rad)
+    node_ordering = get_coordinates_ordering(coords_rad)
 
-    # TODO: AOI mask builder is not used in the current implementation.
+    if aoi_mask_builder is not None:
+        aoi_mask = aoi_mask_builder.get_mask(coords_rad)
+        node_ordering = node_ordering[aoi_mask[node_ordering]]
 
-    nx_graph = create_icosahedral_nx_graph_from_coords(coords_rad, node_ordering)
+    # Creates the graph, with the nodes sorted by latitude and longitude.
+    nx_graph = create_nx_graph_from_tri_coords(coords_rad, node_ordering)
 
     return nx_graph, coords_rad, list(node_ordering)
 
 
-def create_icosahedral_nx_graph_from_coords(coords_rad: np.ndarray, node_ordering: list[int]):
-
+def create_nx_graph_from_tri_coords(coords_rad: np.ndarray, node_ordering: np.ndarray) -> nx.DiGraph:
+    """Creates the networkx graph from the coordinates and the node ordering."""
     graph = nx.DiGraph()
     for i, coords in enumerate(coords_rad[node_ordering]):
         node_id = node_ordering[i]
@@ -57,19 +63,11 @@ def create_icosahedral_nx_graph_from_coords(coords_rad: np.ndarray, node_orderin
     return graph
 
 
-def get_node_ordering(coords_rad: np.ndarray) -> np.ndarray:
-    """Get the node ordering to sort the nodes by latitude and longitude."""
-    # Get indices to sort points by lon & lat in radians.
-    index_latitude = np.argsort(coords_rad[:, 1])
-    index_longitude = np.argsort(coords_rad[index_latitude][:, 0])[::-1]
-    node_ordering = np.arange(coords_rad.shape[0])[index_latitude][index_longitude]
-    return node_ordering
-
-
 def add_edges_to_nx_graph(
     graph: nx.DiGraph,
     resolutions: list[int],
     x_hops: int = 1,
+    area_mask_builder: KNNAreaMaskBuilder | None = None,
 ) -> nx.DiGraph:
     """Adds the edges to the graph.
 
@@ -84,6 +82,8 @@ def add_edges_to_nx_graph(
         Levels of mesh refinement levels to consider.
     x_hops : int, optional
         Number of hops between 2 nodes to consider them neighbours, by default 1.
+    area_mask_builder : KNNAreaMaskBuilder
+        NearestNeighbors with the cloud of points to limit the mesh area, by default None.
 
     Returns
     -------
@@ -94,10 +94,10 @@ def add_edges_to_nx_graph(
 
     sphere = trimesh.creation.icosphere(subdivisions=resolutions[-1], radius=1.0)
     vertices_rad = cartesian_to_latlon_rad(sphere.vertices)
-    node_neighbours = get_neighbours_within_hops(sphere, x_hops)
+    node_neighbours = get_neighbours_within_hops(sphere, x_hops, valid_nodes=list(graph.nodes))
 
     for idx_node, idx_neighbours in node_neighbours.items():
-        add_neigbours_edges(graph, vertices_rad, idx_node, idx_neighbours)
+        add_neigbours_edges(graph, idx_node, idx_neighbours)
 
     tree = BallTree(vertices_rad, metric="haversine")
 
@@ -109,20 +109,25 @@ def add_edges_to_nx_graph(
         # Get the vertices of the isophere
         r_vertices_rad = cartesian_to_latlon_rad(r_sphere.vertices)
 
-        # TODO AOI mask builder is not used in the current implementation.
+        # Limit area of mesh points.
+        if area_mask_builder is not None:
+            aoi_mask = area_mask_builder.get_mask(vertices_rad)
+            valid_nodes = np.where(aoi_mask)[0]
+        else:
+            valid_nodes = None
 
-        node_neighbours = get_neighbours_within_hops(r_sphere, x_hops)
+        node_neighbours = get_neighbours_within_hops(r_sphere, x_hops, valid_nodes=valid_nodes)
 
         _, vertex_mapping_index = tree.query(r_vertices_rad, k=1)
         for idx_node, idx_neighbours in node_neighbours.items():
-            add_neigbours_edges(
-                graph, r_vertices_rad, idx_node, idx_neighbours, vertex_mapping_index=vertex_mapping_index
-            )
+            add_neigbours_edges(graph, idx_node, idx_neighbours, vertex_mapping_index=vertex_mapping_index)
 
     return graph
 
 
-def get_neighbours_within_hops(tri_mesh: trimesh.Trimesh, x_hops: int) -> dict[int, set[int]]:
+def get_neighbours_within_hops(
+    tri_mesh: trimesh.Trimesh, x_hops: int, valid_nodes: list[int] | None = None
+) -> dict[int, set[int]]:
     """Get the neigbour connections in the graph.
 
     Parameters
@@ -131,6 +136,8 @@ def get_neighbours_within_hops(tri_mesh: trimesh.Trimesh, x_hops: int) -> dict[i
         The mesh to consider.
     x_hops : int
         Number of hops between 2 nodes to consider them neighbours.
+    valid_nodes : list[int], optional
+        The list of valid nodes to consider, by default None.
 
     Returns
     -------
@@ -140,10 +147,12 @@ def get_neighbours_within_hops(tri_mesh: trimesh.Trimesh, x_hops: int) -> dict[i
     """
     edges = tri_mesh.edges_unique
 
-    valid_nodes = list(range(len(tri_mesh.vertices)))
+    if valid_nodes is not None:
+        edges = edges[np.isin(tri_mesh.edges_unique, valid_nodes).all(axis=1)]
+    else:
+        valid_nodes = list(range(len(tri_mesh.vertices)))
     graph = nx.from_edgelist(edges)
 
-    # Get a dictionary of the neighbours within 'x_hops' neighbourhood of each node in the graph
     neighbours = {
         i: set(nx.ego_graph(graph, i, radius=x_hops, center=False) if i in graph else []) for i in valid_nodes
     }
@@ -153,7 +162,6 @@ def get_neighbours_within_hops(tri_mesh: trimesh.Trimesh, x_hops: int) -> dict[i
 
 def add_neigbours_edges(
     graph: nx.Graph,
-    vertices: np.ndarray,
     node_idx: int,
     neighbour_indices: Iterable[int],
     self_loops: bool = False,
@@ -165,8 +173,6 @@ def add_neigbours_edges(
     ----------
     graph : nx.Graph
         The graph.
-    vertices : np.ndarray
-        A 2D array of shape (num_vertices, 2) with the planar coordinates of the mesh, in radians.
     node_idx : int
         The node considered.
     neighbour_indices : list[int]
@@ -180,10 +186,6 @@ def add_neigbours_edges(
         if not self_loops and node_idx == neighbour_idx:  # no self-loops
             continue
 
-        location_node = vertices[node_idx]
-        location_neighbour = vertices[neighbour_idx]
-        edge_length = haversine_distances([location_neighbour, location_node])[0][1]
-
         if vertex_mapping_index is not None:
             # Use the same method to add edge in all spheres
             node_neighbour = vertex_mapping_index[neighbour_idx][0]
@@ -191,6 +193,6 @@ def add_neigbours_edges(
         else:
             node, node_neighbour = node_idx, neighbour_idx
 
-        # add edge to the graph (if both source and target nodes are in the graph)
+        # add edge to the graph
         if node in graph and node_neighbour in graph:
-            graph.add_edge(node_neighbour, node, weight=edge_length)
+            graph.add_edge(node_neighbour, node)